+++ app/helpers/adverse_events_helper.rb
@@ -1,0 +1,2 @@
+module AdverseEventsHelper
+end