+++ db/migrate/20150928182442_add_adverse_events_enabled_to_projects.rb
@@ -1,0 +1,5 @@
+class AddAdverseEventsEnabledToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :adverse_events_enabled, :boolean, null: false, default: false
+  end
+end